requirements.txt
Django==3.1.7
django-environ==0.4.5
psycopg2-binary==2.8.6
djangorestframework==3.12.4
django-cors-headers==3.7.0
django-oauth-toolkit==1.5.0
django-rest-framework-social-oauth2==1.1.0